+++ globals.c
@@ -4,7 +4,7 @@
 /* initialized global definitions */
 
 int fillcol = 72;		/* Current fill column          */
-short kbdm[NKBDM];		/* Macro                        */
+int kbdm[NKBDM];		/* Macro                        */
 char *execstr = NULL;		/* pointer to string to execute */
 char golabel[NPAT] = "";	/* current line to go to        */
 int execlevel = 0;		/* execution IF level           */
@@ -63,8 +63,8 @@ struct kill *kbufh = NULL;		/* kill buffer header pointer           */
 int kused = KBLOCK;		/* # of bytes used in kill buffer       */
 struct window *swindow = NULL;	/* saved window pointer                 */
 int cryptflag = FALSE;		/* currently encrypting?                */
-short *kbdptr;			/* current position in keyboard buf */
-short *kbdend = &kbdm[0];	/* ptr to end of the keyboard */
+int *kbdptr;			/* current position in keyboard buf */
+int *kbdend = &kbdm[0];		/* ptr to end of the keyboard */
 int kbdmode = STOP;		/* current keyboard macro mode  */
 int kbdrep = 0;			/* number of repetitions        */
 int restflag = FALSE;		/* restricted use?              */
